+++ src/parser/extract_ambient_light.c
@@ -2,11 +2,10 @@
 #include <stdio.h>
 #include "miniRT.h"
 #include "libft.h"
-#include "settings.h"
 #include "elements.h"
 #include "parser.h"
 
-static int	add_ambient_light(FLOAT *light_ratio, t_vec3 *color, t_app *app)
+static int	add_ambient_light(double *light_ratio, t_vec3 *color, t_app *app)
 {
 	app->active_camera->ambient_intensity = *light_ratio;
 	app->active_camera->ambient = *color;
@@ -16,7 +15,7 @@ static int	add_ambient_light(FLOAT *light_ratio, t_vec3 *color, t_app *app)
 int	extract_ambient_light(const char *line, t_app *app)
 {
 	char	**tokens;
-	FLOAT	light_ratio;
+	double	light_ratio;
 	t_vec3	color;
 
 	tokens = ft_split(line, ' ');
@@ -33,7 +32,7 @@ int	extract_ambient_light(const char *line, t_app *app)
 	light_ratio = ft_atof(tokens[1]);
 	if (errno)
 		return (free_tokens(tokens), -1);
-	if (light_ratio > (FLOAT)1.0 && light_ratio < (FLOAT)0.0)
+	if (light_ratio > (double)1.0 && light_ratio < (double)0.0)
 		return (free_tokens(tokens), -1);
 	color = extract_color(tokens[2]);
 	if (errno)
